Blogs.UnitTests/Core/Markdown/CodeTests.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using Blogs.Core.Markdown;

namespace Blogs.UnitTests.Core.Markdown
{
    [TestClass()]
    public class CodeTests
    {
        Code code = new Code();
        [TestMethod()]
        public void ToHtml_BacktickQuotes_ReturnsCode()
        {
            string input = "Use the `printf()` function.";
            string expected = "<p>Use the <code>printf()</code> function.</p>";
            code.Set(input);
            string actual = code.ToHtml();
            Assert.AreEqual(expected, actual);
        }
        [TestMethod()]
        public void ToHtml_DoubleBacktickQuotes_ReturnsCode()
        {
            string input = "``There is a literal backtick (`) here.``";
            string expected = "<p><code>There is a literal backtick (&grave;) here.</code></p>";
            code.Set(input);
            string actual = code.ToHtml();
            Assert.AreEqual(expected, actual);
        }
        [TestMethod()]
        public void ToHtml_AngleBrackets_ReturnsHtmlCode()
        {
            string input = "Please don't use any `<blink>` tags.";
            string expected = "<p>Please don't use any <code>&lt;blink&gt;</code> tags.</p>";
            code.Set(input);
            string actual = code.ToHtml();
            Assert.AreEqual(expected, actual);
        }
        [TestMethod()]
        public void ToHtml_Ampersands_ReturnsHtmlCode()
        {
            string input = "`&#8212;` is the decimal-encoded equivalent of `&mdash;`.";
            string expected = "<p><code>&amp;#8212;</code> is the decimal-encoded equivalent of <code>&amp;mdash;</code>.</p>";
            code.Set(input);
            string actual = code.ToHtml();
            Assert.AreEqual(expected, actual);
        }
    }
}